refactor: rename `testutil.ts` to `async_value.ts`
